+++ tests/TestClient.py
@@ -3,6 +3,7 @@ import pytest
 import mock
 import faker
 import random
+import datetime
 import sys
 from faker.providers import BaseProvider
 from thrift.transport.TSocket import TSocket
@@ -36,9 +37,14 @@ fake.add_provider(FakerProvider)
 
 
 def range_factory():
-    import datetime
-    param = (datetime.datetime.now(), datetime.datetime.now())
-    return param
+    # import datetime
+    # param = (datetime.datetime.now(), datetime.datetime.now())
+    # return param
+    param = {
+        'start_date': '2019-06-25',
+        'end_date': '2019-06-25'
+    }
+    return Prepare.range(**param)
 
 
 def ranges_factory():
@@ -57,6 +63,12 @@ def table_schema_factory():
 
 def records_factory(dimension):
     return [[random.random() for _ in range(dimension)] for _ in range(20)]
+
+
+def query_ranges_factory():
+    param = [('2019-06-25', '2019-06-25')]
+    return Prepare.ranges(param)
+
 
 
 class TestConnection:
@@ -228,6 +240,20 @@ class TestVector:
         assert res.OK()
         assert isinstance(results, (list, TopKQueryResult))
 
+    @mock.patch.object(Milvus, 'search_vectors')
+    def test_search_vector_with_range(self, search_vectors, client):
+        search_vectors.return_value = Status(), [[ttypes.QueryResult(111, 111)]]
+        param = {
+            'table_name': fake.table_name(),
+            'query_records': records_factory(256),
+            'top_k': random.randint(0, 10),
+            'query_ranges': query_ranges_factory()
+
+        }
+        res, results = client.search_vectors(**param)
+        assert res.OK()
+        assert isinstance(results, (list, TopKQueryResult))
+
     def test_false_vector(self, client):
         param = {
             'table_name': fake.table_name(),
@@ -306,6 +332,18 @@ class TestVector:
         res = client.client_version()
         assert isinstance(res, str)
 
+    @mock.patch.object(MilvusService.Client, 'Ping')
+    def test_server_status(self, Ping, client):
+        res = client.server_status()
+        assert res == 'OK'
+
+        res = client.server_status('abc')
+        assert res == 'OK'
+
+        Ping.return_value = '0.0.0'
+        res = client.server_status('version')
+        assert res == '0.0.0'
+
 
 class TestPrepare:
 
@@ -320,12 +358,12 @@ class TestPrepare:
         assert isinstance(res, ttypes.TableSchema)
 
     def test_range(self):
-        param = range_factory()
-        res = Prepare.range(param[0], param[1])
+        param = {'start_date':'2019-02-02','end_date':'2019-02-02'}
+        res = Prepare.range(**param)
         assert isinstance(res, ttypes.Range)
 
     def test_ranges(self):
-        param = ranges_factory()
+        param = [('2019-02-02','2019-02-02')]
         res = Prepare.ranges(param)
         assert isinstance(res, list)
         assert isinstance(res[0], ttypes.Range)